Allow BindableObserver to be constructed with no internalEmitter.
Constructing with no internalEmitter leaves the BindableObserver in an
incomplete state, where calling any function that relies on events
throws an error.

This was done for the purpose of BindableObservers that would be
constructed with EventEmitters that are only obtained after, for
example, one is given in a callback. Specifically, this change is
necessary to create a child class that automatically uses a socket
instance obtained from WebSocket or SocketIO.